+++ packages/stream_chat_flutter/lib/src/message_reactions_modal.dart
@@ -236,6 +236,8 @@ class MessageReactionsModal extends StatelessWidget {
             reaction.user!.name.split(' ')[0],
             style: chatThemeData.textTheme.footnoteBold,
             textAlign: TextAlign.center,
+            overflow: TextOverflow.ellipsis,
+            maxLines: 1,
           ),
         ],
       ),