+++ packages/stream_chat_flutter/lib/src/utils.dart
@@ -8,9 +8,9 @@ import 'package:url_launcher/url_launcher.dart';
 
 /// Launch URL
 Future<void> launchURL(BuildContext context, String url) async {
-  if (await canLaunch(url)) {
-    await launch(url);
-  } else {
+  try {
+    await launch(Uri.parse(url).withScheme.toString());
+  } catch (e) {
     ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(content: Text(context.translations.launchUrlError)),
     );